refactor(glsl): move the varying split out of the translator
A matrix, a struct and an array varying are one reflected name over several
locations, and all three are answered the same way: the declaration comes
out of the body, one varying per column, member or element goes into the
header, and the pack's own name is rebuilt around its main. Those three
passes, the three lists they fill, the naming of what they emit and the six
prologues and epilogues that copy the values now live in VaryingSplit. It
reads the tokens, the unit's liveness and the stage, and takes the reader
for a file scope declaration from the translator rather than keeping a
second one of its own.
